reworked navigation and added statistics screen

+++ feature/word/src/main/java/com/example/word/screen/main/WordScreen.kt
@@ -1,24 +1,36 @@
 package com.example.word.screen.main
 
-import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.Warning
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.domain.model.WordModel
 import com.example.ui.AddButton
 import com.example.ui.SearchInputView
 import com.example.ui.WordsList
+import com.example.word.R
 import org.koin.androidx.compose.koinViewModel
-
 
 @Composable
 fun WordScreen(
@@ -32,17 +44,25 @@ fun WordScreen(
 
     val state by viewModel.state
 
+    val isOpenDialog = remember {
+        mutableStateOf(false)
+    }
+
+    var selectedWord by remember {
+        mutableStateOf<WordModel?>(null)
+    }
+
 
     Box(
         modifier = Modifier
-//            .background(Color.White)
-
             .fillMaxSize()
     ) {
 
         Column(
             modifier = Modifier
-                .padding(horizontal = 15.dp)
+                .padding(horizontal = 15.dp),
+            verticalArrangement = Arrangement.Top,
+            horizontalAlignment = Alignment.CenterHorizontally
         ) {
 
             SearchInputView(
@@ -53,19 +73,56 @@ fun WordScreen(
                 viewModel.handleIntent(WordScreenIntent.SearchWord(state.searchReq))
             }
 
-            if (state.isError) {
-                Text(text = state.errorMessage)
-            }
-
-            if (state.isLoading) {
-                Text(text = "Loading")
-            }
-
-            if(!state.isError && !state.isLoading) {
-                WordsList(words = state.words) {
-                    viewModel.handleIntent(intent = WordScreenIntent.DeleteWord(it))
+            when {
+                state.isLoading -> Text(text = stringResource(id = R.string.loading))
+                state.isError -> Text(text = state.errorMessage)
+                else -> {
+                    WordsList(words = state.words) {
+                        isOpenDialog.value = true
+                        selectedWord = it
+                    }
                 }
             }
+
+            if (isOpenDialog.value) {
+                AlertDialog(
+                    onDismissRequest = {
+                        isOpenDialog.value = false
+                    },
+                    confirmButton = {
+                        TextButton(onClick = {
+                            isOpenDialog.value = false
+                            selectedWord?.let {
+                                viewModel.handleIntent(intent = WordScreenIntent.DeleteWord(it))
+                            }
+                        }) {
+                            Text(text = "Delete", color = Color.Red)
+                        }
+                    },
+                    dismissButton = {
+                        TextButton(onClick = {
+                            isOpenDialog.value = false
+                        }) {
+                            Text(text = "Close")
+                        }
+                    },
+                    title = { Text(text = stringResource(id = R.string.delete_alert_title)) },
+                    text = {
+                        Text(
+                            text = stringResource(id = R.string.delete_word_alert_text),
+                            fontSize = 20.sp,
+                            fontWeight = FontWeight.SemiBold
+                        )
+                    },
+                    icon = {
+                        Icon(
+                            imageVector = Icons.Rounded.Warning,
+                            contentDescription = "Warning icon"
+                        )
+                    }
+                )
+            }
+
 
         }
